pyproject.toml
[project]
name = "cc-wait"
version = "0.2.0"
description = "Automatically continue Claude Code sessions after rate limits reset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-fasthtml>=0.12.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
]

[project.scripts]
cc-wait = "cc_wait.cli:main"
cc-wait-server = "cc_wait.server:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.0",
    "ty>=0.0.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["cc_wait"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
